pyproject.toml
[tool.poetry]
name = "geo_ii"
version = "0.1.0"
description = ""
authors = ["Manul <naivze@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
matplotlib = "^3.7.1"
tensorflow = "^2.11.0"
tuya-iot-py-sdk = "^0.6.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
